Report logika izmestena u ReportService
Napravljena 3 odvojena DTO za reportove umesto jednog
