+++ backend/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 #&serverTimezone=UTC
 spring.datasource.url=jdbc:mysql://localhost:3306/Ticketist?useSSL=false&createDatabaseIfNotExist=true
-spring.datasource.username=hbstudent
-spring.datasource.password=hbstudent
+spring.datasource.username=root
+spring.datasource.password=admin
 #spring.datasource.initialization-mode=always
 
 spring.jpa.show-sql=true
